Fix encryption and decryption for padded files
The decryption code assumed that the raw and stored size is the same for
uncompressed files. This was fine for previously tested games, but not
really for "Mamoru-kun wa Norowarete Shimatta!" (Xbox 360). That game's
archives *do* contain uncompressed files with different raw and stored
sizes. This difference is the result of the reference implementation
zero-padding uncompressed files to a multiple of 64 bytes.

Although none of the padded files ended up being padded to a multiple of
256 bytes, further investigation revealed that the encryption "key" is
derived from a file's *raw* size, not its stored size. This commit
corrects the decryption code to derive the "key" from the raw size and
remove the padding after decryption.

When writing uncompressed files, the encryption code pads the file to a
multiple of 64 bytes before encryption. The padding doesn't appear to be
necessary to allow games to correctly read uncompressed files, but the
reference implementation did it, so ggxtool shall do the same!