Decompress lazily, fix XOR special case
Decompression is now only performed when actually needed, such as when
extracting files from archives. Compression is performed when adding
files to archives. Because of this change, the ArchiveFile type now also
stores a file's raw size in addition to its stored size (implicitly, in
the Data array's length). When loading or saving archives, only the XOR
is applied because it depends on the file's location within the archive.

Additionally, files where (rawSize / 256 + 5) % 256 == 0 are now handled
correctly, using 0x05 instead of 0x00 to mask the key block. This change
matches the implementation by @dakrk.

For testing, a new "decompress" command can be used to decompress files
within the archive.